refactor(silencer/ui-v2): extract OptionsControlsRuntime
Moves Game::RenderOptionsControlsV2 / DispatchOptionsControlsV2Click
/ TickOptionsControlsV2 + Game::StartControlsRebind + the file-static
controls helpers + rebind capture state machine into
OptionsControlsRuntime co-located with BuildOptionsControls in
screens/options_controls.{h,cpp}.

DispatchKeyDown latches the next scancode while a slot is armed;
Tick() commits + clears. Events.cpp's keyboard branch now delegates
to active_runtime->DispatchKeyDown unconditionally; the dedicated
OPTIONSCONTROLS branch is gone.

Removed from Game: 5 controls_rebind_* member fields,
StartControlsRebind public method, RenderOptionsControlsV2 /
DispatchOptionsControlsV2Click / TickOptionsControlsV2.

15-screen PPM byte-identity verified.
